+++ test/setup.js
@@ -1,2 +1,1 @@
 global.DEBUG = false;
-global.Promise = require('bluebird');