+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0] - TBD
 
+- Added external `rs.config.js` manifest support
 - Changed the main control flow from `async` to Bluebird Promise.
 - Removed `async` dependency