+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - TBD
+
+- Removed global usage for `Bluebird` dependency
+
 ## [0.2.0] - 2018-03-10
 
 - Added external `rs.config.js` manifest support